格式化项目

+++ src/main/java/top/guoziyang/mydb/backend/utils/Types.java
@@ -2,8 +2,8 @@ package top.guoziyang.mydb.backend.utils;
 
 public class Types {
     public static long addressToUid(int pgno, short offset) {
-        long u0 = (long)pgno;
-        long u1 = (long)offset;
+        long u0 = pgno;
+        long u1 = offset;
         return u0 << 32 | u1;
     }
 }
